+++ packages/xcm-cfg/src/builders/extrinsics/balances.ts
@@ -1,0 +1,26 @@
+import {
+  ExtrinsicConfig,
+  ExtrinsicConfigBuilder,
+} from '@galacticcouncil/xcm-core';
+
+const pallet = 'balances';
+
+const transferAll = (): ExtrinsicConfigBuilder => {
+  const func = 'transferAll';
+  return {
+    build: ({ address }) =>
+      new ExtrinsicConfig({
+        module: pallet,
+        func,
+        getArgs: () => {
+          return [address, true];
+        },
+      }),
+  };
+};
+
+export const balances = () => {
+  return {
+    transferAll,
+  };
+};