+++ packages/xcm-cfg/src/configs/kusama/kusama.ts
@@ -4,6 +4,8 @@ import { ksm } from '../../assets';
 import { kusama, kusamaAssetHub, basilisk, karura } from '../../chains';
 import { BalanceBuilder, ExtrinsicBuilder } from '../../builders';
 
+const xcmDeliveryFee = 0.002;
+
 const toBasilisk: AssetRoute[] = [
   new AssetRoute({
     source: {
@@ -12,6 +14,7 @@ const toBasilisk: AssetRoute[] = [
       fee: {
         asset: ksm,
         balance: BalanceBuilder().substrate().system().account(),
+        extra: xcmDeliveryFee,
       },
       destinationFee: {
         balance: BalanceBuilder().substrate().system().account(),
@@ -37,6 +40,7 @@ const toKarura: AssetRoute[] = [
       fee: {
         asset: ksm,
         balance: BalanceBuilder().substrate().system().account(),
+        extra: xcmDeliveryFee,
       },
       destinationFee: {
         balance: BalanceBuilder().substrate().system().account(),
@@ -62,6 +66,7 @@ const toAssetHub: AssetRoute[] = [
       fee: {
         asset: ksm,
         balance: BalanceBuilder().substrate().system().account(),
+        extra: xcmDeliveryFee,
       },
       destinationFee: {
         balance: BalanceBuilder().substrate().system().account(),