+++ packages/xcm-cfg/src/configs/polkadot/hydration/templates.ts
@@ -14,6 +14,7 @@ import {
 } from '../../../builders';
 import {
   assetHub,
+  assetHubCex,
   ethereum,
   moonbeam,
   solana,
@@ -23,8 +24,10 @@ import { Tag } from '../../../tags';
 
 import { balance, fee } from './configs';
 
-export const MRL_EXECUTION_FEE = 0.9;
-export const MRL_XCM_FEE = 1;
+export const MRL_EXECUTION_FEE = 0.9; // Remote execution fee (< 0.9)
+export const MRL_XCM_FEE = 1; // Destination fee (< 0.1) + Remote execution fee (< 0.9)
+
+export const CEX_EXECUTION_FEE = 0.02; // Remote execution fee (< 0.02)
 
 const isDestinationFeeSwapSupported = (
   params: ExtrinsicConfigBuilderParams
@@ -34,9 +37,7 @@ const isDestinationFeeSwapSupported = (
   return !!enabled;
 };
 
-const getSwapExtrinsicBuilder = () => {
-  return ExtrinsicBuilder().router().buy({ withSlippage: 30 });
-};
+const swapExtrinsicBuilder = ExtrinsicBuilder().router().buy({ slippage: 30 });
 
 export function toHubExtTemplate(asset: Asset): AssetRoute {
   return new AssetRoute({
@@ -58,7 +59,7 @@ export function toHubExtTemplate(asset: Asset): AssetRoute {
     },
     extrinsic: ExtrinsicDecorator(
       isDestinationFeeSwapSupported,
-      getSwapExtrinsicBuilder()
+      swapExtrinsicBuilder
     ).prior(ExtrinsicBuilder().xTokens().transferMultiassets()),
   });
 }
@@ -83,7 +84,7 @@ export function toMoonbeamErc20Template(asset: Asset): AssetRoute {
     },
     extrinsic: ExtrinsicDecorator(
       isDestinationFeeSwapSupported,
-      getSwapExtrinsicBuilder()
+      swapExtrinsicBuilder
     ).prior(ExtrinsicBuilder().xTokens().transferMultiCurrencies()),
   });
 }
@@ -108,8 +109,79 @@ export function toZeitgeistErc20Template(asset: Asset): AssetRoute {
     },
     extrinsic: ExtrinsicDecorator(
       isDestinationFeeSwapSupported,
-      getSwapExtrinsicBuilder()
+      swapExtrinsicBuilder
     ).prior(ExtrinsicBuilder().xTokens().transferMultiCurrencies()),
+  });
+}
+
+export function toCexViaAssethubTemplate(asset: Asset): AssetRoute {
+  return new AssetRoute({
+    source: {
+      asset: asset,
+      balance: balance(),
+      fee: fee(),
+      destinationFee: {
+        balance: balance(),
+      },
+    },
+    destination: {
+      chain: assetHubCex,
+      asset: asset,
+      fee: {
+        amount: FeeAmountBuilder().Assethub().destFeeIn(asset),
+        asset: asset,
+      },
+    },
+    extrinsic: ExtrinsicBuilder()
+      .utility()
+      .batchAll([
+        ExtrinsicBuilder().polkadotXcm().transferAssetsUsingTypeAndThen({
+          transferType: XcmTransferType.DestinationReserve,
+        }),
+        ExtrinsicBuilder().polkadotXcm().send().transferAsset({
+          fee: CEX_EXECUTION_FEE,
+        }),
+      ]),
+  });
+}
+
+export function toCexViaAssethubTemplate2(asset: Asset): AssetRoute {
+  return new AssetRoute({
+    source: {
+      asset: asset,
+      balance: balance(),
+      fee: fee(),
+      destinationFee: {
+        balance: balance(),
+      },
+    },
+    destination: {
+      chain: assetHubCex,
+      asset: asset,
+      fee: {
+        amount: FeeAmountBuilder().Assethub().destFeeIn(asset),
+        asset: asset,
+      },
+    },
+    extrinsic: ExtrinsicBuilder()
+      .utility()
+      .batchAll([
+        ExtrinsicBuilder().polkadotXcm().transferAssetsUsingTypeAndThen({
+          transferType: XcmTransferType.DestinationReserve,
+        }),
+        ExtrinsicBuilder().polkadotXcm().send().transact({
+          fee: CEX_EXECUTION_FEE,
+        }),
+      ]),
+    transact: {
+      chain: assetHub,
+      fee: {
+        amount: 0,
+        asset: asset,
+        balance: balance(),
+      },
+      extrinsic: ExtrinsicBuilder().assets().transfer(),
+    },
   });
 }
 
@@ -140,7 +212,7 @@ export function toEthereumViaWormholeTemplate(
     },
     extrinsic: ExtrinsicDecorator(
       isDestinationFeeSwapSupported,
-      getSwapExtrinsicBuilder()
+      swapExtrinsicBuilder
     ).priorMulti([
       ExtrinsicBuilder().xTokens().transferMultiCurrencies(),
       ExtrinsicBuilder().polkadotXcm().send().transact({
@@ -227,7 +299,7 @@ export function toSolanaViaWormholeTemplate(
     },
     extrinsic: ExtrinsicDecorator(
       isDestinationFeeSwapSupported,
-      getSwapExtrinsicBuilder()
+      swapExtrinsicBuilder
     ).priorMulti([
       ExtrinsicBuilder().xTokens().transferMultiCurrencies(),
       ExtrinsicBuilder().polkadotXcm().send().transact({