+++ packages/xcm-sdk/src/platforms/substrate/index.ts
@@ -1,2 +1,4 @@
 export * from './SubstratePlatform';
 export * from './SubstrateService';
+
+export * from './types';