+++ packages/xcm-sdk/src/platforms/substrate/types.ts
@@ -1,0 +1,11 @@
+import { Asset } from '@galacticcouncil/xcm-core';
+
+import { Call } from '../types';
+
+export interface SubstrateCall extends Call {
+  txOptions:
+    | {
+        asset: Asset | undefined;
+      }
+    | undefined;
+}