+++ packages/xcm-core/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @galacticcouncil/xcm-core
 
+## 6.1.0
+
+### Minor Changes
+
+- Dex re-work
+- Chain currency asset
+- Parachain flags:
+  - treasury
+  - cex forwarding
+  - delivery fee
+
 ## 6.0.0
 
 ### Major Changes